Move package subcommands -> child
